LayoutTests/imported/w3c/web-platform-tests/html/canvas/offscreen/drawing-rectangles-to-the-canvas/2d.fillRect.clip.worker-expected.txt


PASS fillRect is affected by clipping regions 
